bluetooth/requirements.txt
dbus-python>=1.3.2
bleak>=0.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
